// src/types.ts
export interface StationData {
  "Company Name": string;
  "Store Name": string;
  "City": string;
  "Street Address": string;
  "Country": string;
  "Title": string;
  "Zipcode": string;
  "Address": string;
  "State Full": string;
  "Phone": string;
  "Hours": string;
  "Web": string;
  "Latitude": number;
  "Longitude": number;
  "Warehouse": string;
  "Regular": string;
  "Premium": string;
  "Diesel": string;
  "Gas Price Types": string;
  "Last Updated": string;
  "Disclaimer": string;
  "post_author": string;
}